fix: avoid panic

+++ crates/asm/txs/bridge-v1/src/deposit/mod.rs
@@ -39,7 +39,7 @@
 mod parse;
 mod validation;
 
-pub const DEPOSIT_OUTPUT_INDEX: u32 = 1;
+pub const DEPOSIT_OUTPUT_INDEX: usize = 1;
 
 pub use parse::{DepositInfo, MIN_DEPOSIT_TX_AUX_DATA_LEN, parse_deposit_tx};
 pub use validation::{validate_deposit_output_lock, validate_drt_spending_signature};
